+++ src/Connection.php
@@ -81,12 +81,11 @@ class Connection implements ConnectionInterface
     /**
      * @param string|int $id
      */
-    public function delete($id): self
+    public function delete($id): void
     {
         $id = $this->filterJobId($id);
         (new Command\Delete($id))
             ->process($this->socket);
-        return $this;
     }
 
     /**
@@ -96,33 +95,30 @@ class Connection implements ConnectionInterface
         $id,
         int $priority = ConnectionInterface::DEFAULT_PRIORITY,
         int $delay = ConnectionInterface::DEFAULT_DELAY
-    ): self {
+    ): void {
         $id = $this->filterJobId($id);
         (new Command\Release($id, $priority, $delay))
             ->process($this->socket);
-        return $this;
     }
 
     /**
      * @param string|int $id
      */
-    public function bury($id, int $priority = ConnectionInterface::DEFAULT_PRIORITY): self
+    public function bury($id, int $priority = ConnectionInterface::DEFAULT_PRIORITY): void
     {
         $id = $this->filterJobId($id);
         (new Command\Bury($id, $priority))
             ->process($this->socket);
-        return $this;
     }
 
     /**
      * @param string|int $id
      */
-    public function touch($id): self
+    public function touch($id): void
     {
         $id = $this->filterJobId($id);
         (new Command\Touch($id))
             ->process($this->socket);
-        return $this;
     }
 
     public function watch(string $tube): int