+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Update commands to better reflect the protocol:
   - `watch()` returns integer of number of watched tubes.
   - `ignore()` throws `CommandException` if trying to ignore the final tube.
-  - `useTube()` has void return type.
+  - `useTube()`, `touch()`, `release()`, `bury()`, `delete()`
+    have void return type.
 - Status is a required argument for the `tube:peek` CLI command.
 - **BC break**: Constructor for `Connection` no longer needs a `Socket`.
   Pass the same parameters directly to `Connection`.